# doc/swagger/brood.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "brood API",
    "version": "1.0",
    "contact": {
      "name": "itsscb",
      "url": "https://github.com/itsscb/brood",
      "email": "dev@itsscb.de"
    }
  },
  "tags": [
    {
      "name": "brood",
      "description": "brood service"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/join": {
      "post": {
        "operationId": "brood_Join",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/pbSpore"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of pbSpore"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbJoinBrood"
            }
          }
        ],
        "tags": [
          "brood"
        ]
      }
    },
    "/v1/send_spore": {
      "post": {
        "operationId": "brood_SendSpore",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbAcknowledgeSpore"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbSpore"
            }
          }
        ],
        "tags": [
          "brood"
        ]
      }
    }
  },
  "definitions": {
    "pbAcknowledgeSpore": {
      "type": "object",
      "properties": {
        "ack": {
          "type": "boolean"
        }
      },
      "title": "AcknowledgeSpore"
    },
    "pbJoinBrood": {
      "type": "object",
      "properties": {
        "hive": {
          "type": "string"
        },
        "topics": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "title": "JoinBrood"
    },
    "pbSpore": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "timestamp": {
          "type": "string",
          "format": "date-time",
          "example": "2023-12-10T00:00:00Z"
        },
        "topic": {
          "type": "string"
        },
        "data": {
          "type": "string"
        }
      },
      "title": "Spore"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}